+++ src/test/resources/com/pestphp/pest/inspections/NoDuplicateCustomExpectation.php
@@ -1,0 +1,9 @@
+<?php
+
+expect()->extend('toBeOne', function (): bool {
+    return true;
+});
+
+expect()->extend('toBeTwo', function (): bool {
+    return true;
+});